bump version to v3.2.4barrykn1 (58.21114.1)
I feel the need to do something to the version number to distinguish it
from upstream, in case I end up deploying this on my own systems (or in
case any unforeseen nightmare scenario happens). So, here goes...

(Explanation of 21114: B is 2nd letter of alphabet, K is 11th, N is 14th.
Those are my initials. Concatenate 2, 11, 14.)

I tested a similar patch recently, but not this exact patch, so if this
somehow breaks the build, then at that point I'll revert it and do
something different.